@@ -1,12 +1,18 @@
 module blog.rtfb.lt/migrate-db
 
-go 1.16
+go 1.18
 
 require (
 	github.com/docopt/docopt-go v0.0.0-20180111231733-ee0de3bc6815
 	github.com/golang-migrate/migrate/v4 v4.15.2
 	github.com/lib/pq v1.10.2
 	github.com/rtfb/bark v0.0.0-20150422192901-07d6537e2e59
-	github.com/rtfb/httputil v0.0.0-20150217190924-9649b2ef6634 // indirect
 	gopkg.in/yaml.v2 v2.4.0
+)
+
+require (
+	github.com/hashicorp/errwrap v1.1.0 // indirect
+	github.com/hashicorp/go-multierror v1.1.1 // indirect
+	github.com/rtfb/httputil v0.0.0-20150217190924-9649b2ef6634 // indirect
+	go.uber.org/atomic v1.7.0 // indirect
 )
